# specs/001-primoinspect-digital-inspection/contracts/api-spec.yaml
openapi: 3.0.3
info:
  title: PrimoInspect Digital Inspection Platform API
  description: REST API for renewable energy inspection management with mobile-first design
  version: 1.0.0
  contact:
    name: PrimoInspect API Support
    email: api@primoinspect.com
servers:
  - url: https://api.primoinspect.com/v1
    description: Production server
  - url: http://localhost:3000/api
    description: Development server
paths:
...
  /projects/{projectId}:
    get:
      summary: Get project details
      tags: [Projects]
      security:
        - sessionAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /inspections:
    get:
      summary: List inspections with filters
      tags: [Inspections]
      security:
        - sessionAuth: []
      parameters:
        - name: projectId
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, pending, in-review, approved, rejected]
        - name: assignedTo
          in: query
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
      responses:
        "200":
          description: List of inspections
          content:
            application/json:
              schema:
                type: object
                properties:
                  inspections:
                    type: array
                    items:
                      $ref: "#/components/schemas/Inspection"
                  pagination:
                    $ref: "#/components/schemas/Pagination"

    post:
      summary: Create inspection assignment
      tags: [Inspections]
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInspection"
      responses:
        "201":
          description: Inspection created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inspection"
        "400":
          $ref: "#/components/responses/BadRequest"
  /inspections/{inspectionId}:
    get:
      summary: Get inspection details with evidence
      tags: [Inspections]
      security:
        - sessionAuth: []
      parameters:
        - name: inspectionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Inspection with evidence and timeline
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InspectionDetail"

    put:
      summary: Update inspection (submit responses)
      tags: [Inspections]
      security:
        - sessionAuth: []
      parameters:
        - name: inspectionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInspection"
      responses:
        "200":
          description: Inspection updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inspection"
  /inspections/{inspectionId}/approve:
    post:
      summary: Approve or reject inspection
      tags: [Approvals]
      security:
        - sessionAuth: []
      parameters:
        - name: inspectionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [decision, notes]
              properties:
                decision:
                  type: string
                  enum: [approved, rejected]
                notes:
                  type: string
                  minLength: 1
                  maxLength: 1000
      responses:
        "200":
          description: Approval recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Approval"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: Inspection already approved/rejected

  # Evidence Management
...
  /notifications:
    get:
      summary: Get user notifications
      tags: [Notifications]
      security:
        - sessionAuth: []
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
            default: false
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
      responses:
        "200":
          description: List of notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"

    patch:
      summary: Mark notifications as read
      tags: [Notifications]
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [notificationIds]
              properties:
                notificationIds:
                  type: array
                  items:
                    type: string
                    format: uuid
      responses:
        "200":
          description: Notifications marked as read
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
  schemas:
    Project:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 100
        description:
          type: string
        status:
          type: string
          enum: [active, completed, on-hold, cancelled]
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        location:
          $ref: "#/components/schemas/Location"
        teamMembers:
          type: array
          items:
            type: string
            format: uuid
          maxItems: 10
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Inspection:
      type: object
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        checklistId:
          type: string
          format: uuid
        assignedTo:
          type: string
          format: uuid
        title:
          type: string
          maxLength: 200
        description:
          type: string
        status:
          type: string
          enum: [draft, pending, in-review, approved, rejected]
        priority:
          type: string
          enum: [low, medium, high]
        dueDate:
          type: string
          format: date
        location:
          $ref: "#/components/schemas/Location"
        rejectionCount:
          type: integer
          minimum: 0
          maximum: 2
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        submittedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
    InspectionDetail:
      allOf:
        - $ref: "#/components/schemas/Inspection"
        - type: object
          properties:
            responses:
              type: object
            evidence:
              type: array
              items:
                $ref: "#/components/schemas/Evidence"
            approvals:
              type: array
              items:
                $ref: "#/components/schemas/Approval"
            timeline:
              type: array
              items:
                $ref: "#/components/schemas/TimelineEvent"
    CreateInspection:
      type: object
      required: [projectId, checklistId, assignedTo, title]
      properties:
        projectId:
          type: string
          format: uuid
        checklistId:
          type: string
          format: uuid
        assignedTo:
          type: string
          format: uuid
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
        priority:
          type: string
          enum: [low, medium, high]
          default: medium
        dueDate:
          type: string
          format: date
    UpdateInspection:
      type: object
      properties:
        responses:
          type: object
        location:
          $ref: "#/components/schemas/Location"
        status:
          type: string
          enum: [draft, pending]
    Evidence:
      type: object
      properties:
        id:
          type: string
          format: uuid
        inspectionId:
          type: string
          format: uuid
        uploadedBy:
          type: string
          format: uuid
        filename:
          type: string
        originalName:
          type: string
        mimeType:
          type: string
        fileSize:
          type: integer
          maximum: 52428800 # 50MB
        url:
          type: string
          format: uri
        thumbnailUrl:
          type: string
          format: uri
        location:
          $ref: "#/components/schemas/Location"
        timestamp:
          type: string
          format: date-time
        verified:
          type: boolean
        annotations:
          type: array
          items:
            type: object
            properties:
              x:
                type: number
              y:
                type: number
              text:
                type: string
        createdAt:
          type: string
          format: date-time
    Approval:
      type: object
      properties:
        id:
          type: string
          format: uuid
        inspectionId:
          type: string
          format: uuid
        approverId:
          type: string
          format: uuid
        decision:
          type: string
          enum: [approved, rejected]
        notes:
          type: string
          minLength: 1
          maxLength: 1000
        reviewDate:
          type: string
          format: date-time
        isEscalated:
          type: boolean
        escalationReason:
          type: string
        previousApprovalId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
...
    Notification:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        type:
          type: string
          enum:
            [
              assignment,
              status-change,
              approval-required,
              escalation,
              report-ready,
            ]
        title:
          type: string
          maxLength: 100
        message:
          type: string
          maxLength: 500
        relatedEntityType:
          type: string
          enum: [inspection, project, approval, report]
        relatedEntityId:
          type: string
          format: uuid
        isRead:
          type: boolean
        priority:
          type: string
          enum: [low, medium, high]
        createdAt:
          type: string
          format: date-time
...
    TimelineEvent:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [created, submitted, approved, rejected, escalated]
        description:
          type: string
        userId:
          type: string
          format: uuid
        userName:
          type: string
        timestamp:
          type: string
          format: date-time
        metadata:
          type: object
    Location:
      type: object
      required: [latitude, longitude]
      properties:
        latitude:
          type: number
          minimum: -90
          maximum: 90
        longitude:
          type: number
          minimum: -180
          maximum: 180
        accuracy:
          type: number
          minimum: 0
        address:
          type: string
    Pagination:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
          maximum: 50
        total:
          type: integer
          minimum: 0
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string
        details:
          type: object
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...